Refactor CustomUserRepositoryImpl to use GenericRepositoryWithPagination
